Move Python files into package directory
