--- config.py
@@ -1,22 +1,0 @@
-# This file is part of SNNAP2NEURON.
-#
-# Copyright (C) 2019 Jayalath A M M Abeywardhana, Jeffrey Gill, Reid Bolding,
-# Peter Thomas
-#
-# SNNAP2NEURON is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# any later version.
-#
-# SNNAP2NEURON is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with SNNAP2NEURON.  If not, see <https://www.gnu.org/licenses/>.
-class config():
-    """
-    
-    """
-    pass
