+++ snnap2neuron/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""
+An automated translation tool for converting SNNAP models into NEURON models
+"""
