+++ snnap2neuron/network.py
