+++ snnap2neuron/readS.py
