+++ snnap2neuron/util.py
